create templates controller;
